refs. #BE-6: AWS CRT-HTTP-Client in die DynamoDB-Konfiguration integriert und fetchArticles in DummyDataFactory umgeschrieben.
